requirements.txt
# Core dependencies for CXR-Assistant
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
